Address review comment - move new function to SAC file
